// libs/supabase/schema.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
      gists: {
        Row: {
          content: string
          created_at: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_id: string | null
          title: string
        }
        Insert: {
          content: string
          created_at?: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_id?: string | null
          title: string
        }
        Update: {
          content?: string
          created_at?: string
          description?: string
          id?: string
          is_paid?: boolean
          lang?: string
          price?: number
          profile_id?: string | null
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "gists_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          address: Json | null
          avatar_url: string
          bio: string | null
          created_at: string
          email: string
          id: string
          name: string
          payment_connected_account: string | null
          phone: string | null
          site: string | null
          username: string
        }
        Insert: {
          address?: Json | null
          avatar_url: string
          bio?: string | null
          created_at?: string
          email: string
          id: string
          name: string
          payment_connected_account?: string | null
          phone?: string | null
          site?: string | null
          username: string
        }
        Update: {
          address?: Json | null
          avatar_url?: string
          bio?: string | null
          created_at?: string
          email?: string
          id?: string
          name?: string
          payment_connected_account?: string | null
          phone?: string | null
          site?: string | null
          username?: string
        }
        Relationships: [
          {
            foreignKeyName: "profiles_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      sales: {
        Row: {
          created_at: string
          customer_email: string
          gist_id: string | null
          id: string
        }
        Insert: {
          created_at?: string
          customer_email: string
          gist_id?: string | null
          id: string
        }
        Update: {
          created_at?: string
          customer_email?: string
          gist_id?: string | null
          id?: string
        }
        Relationships: [
          {
            foreignKeyName: "sales_gist_id_fkey"
            columns: ["gist_id"]
            isOneToOne: false
            referencedRelation: "gists"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never